moved MatchRestEndpoint tests to separate package + added deleteMatchTest()
